+++ registrations/resolvers.js
@@ -38,6 +38,91 @@ const resolvers = {
     id(registration, _args, _context, _info) {
       return registration._id.toString()
     },
+    // Retrieve the progress of the user associated to this registration.
+    async progress(registration, _args, { models }, _info) {
+      const { Competency, Course, ProgressHistory } = models
+
+      // Retrieve the course associated to this registration.
+      const course = await Course.findOne(
+        { _id: registration.course },
+        'competencies'
+      ).lean()
+      if (!course) {
+        throw new UserInputError('COURSE_NOT_FOUND')
+      }
+
+      // Retrieve the progress histories for each competency.
+      const competencies = await Promise.all(
+        course.competencies.map(async (c) => ({
+          category: c.category,
+          competency: await Competency.findOne({ _id: c.competency }).lean(),
+          history: await ProgressHistory.find(
+            {
+              competency: c.competency,
+              user: registration.user,
+            },
+            'learningOutcomes stars'
+          ).lean(),
+          useLearningOutcomes: !!c.useLearningOutcomes,
+        }))
+      )
+
+      // Computes the progress for each competency.
+      competencies.forEach((competency) => {
+        // If stars are used, compute the total number of acquired stars.
+        if (!competency.useLearningOutcomes) {
+          competency.stars = competency.history.reduce(
+            (stars, c) => stars + c.stars,
+            0
+          )
+
+          competency.innerProgress = Math.min(competency.stars, 5)
+          competency.progress = competency.innerProgress
+        }
+        // If learning outcomes are used, compute the total number
+        // of acquired learning outcomes, for each one of them
+        else {
+          competency.learningOutcomes = competency.history.reduce(
+            (learningOutcomes, c) => {
+              c.learningOutcomes.forEach((i) => (learningOutcomes[i] += 1))
+              return learningOutcomes
+            },
+            Array.from(
+              { length: competency.competency.learningOutcomes?.length },
+              () => 0
+            )
+          )
+
+          const sumAcquired = competency.learningOutcomes.reduce(
+            (acc, n) => acc + n,
+            0
+          )
+          const totalToAcquire = competency.competency.learningOutcomes.reduce(
+            (acc, lo) => acc + (lo.takes ?? 1),
+            0
+          )
+          competency.innerProgress = (5 * sumAcquired) / totalToAcquire
+          competency.progress = Math.trunc(20 * competency.innerProgress)
+        }
+      })
+
+      const basic = competencies.filter((c) => c.category === 'basic')
+      const advanced = competencies.filter((c) => c.category === 'advanced')
+
+      return {
+        advanced: Math.trunc(
+          (advanced.reduce((acc, c) => acc + c.innerProgress, 0) /
+            (advanced.length * 5)) *
+            100
+        ),
+        basic: Math.trunc(
+          (basic.reduce((acc, c) => acc + c.innerProgress, 0) /
+            (basic.length * 5)) *
+            100
+        ),
+        competencies,
+      }
+    },
   },
   Query: {
     async registration(_parent, args, { models, user }, _info) {
@@ -295,6 +380,54 @@ const resolvers = {
 
       return null
     },
+    // Delete a registration.
+    async deleteRegistration(_parent, args, { models, user }, _info) {
+      const { Course, Registration } = models
+
+      // Retrieve the registration to delete.
+      const registration = await Registration.findOne(
+        { _id: args.id },
+        'course'
+      )
+      if (!registration) {
+        throw new UserInputError('REGISTRATION_NOT_FOUND')
+      }
+
+      // Retrieve the course associated to the registration.
+      const course = await Course.findOne(
+        { _id: registration.course },
+        'archived coordinator published'
+      ).lean()
+      if (!course) {
+        throw new UserInputError('COURSE_NOT_FOUND')
+      }
+
+      // Only the coordinator can delete a registration
+      // for a course with the 'published' status.
+      if (
+        !isCoordinator(course, user) ||
+        !course.published ||
+        course.archived
+      ) {
+        throw new UserInputError('REGISTRATION_DELETE')
+      }
+
+      // If the registration is confirmed, need to delete all the information
+      // associated to the learner and related to the registration
+      if (!registration.invitation) {
+        // TODO: delete everything related to the confirmed registration
+        throw new UserInputError('UPCOMING_FEATURE')
+      }
+
+      try {
+        await registration.delete()
+        return true
+      } catch (err) {
+        Bugsnag.notify(err)
+      }
+
+      return false
+    },
     // Register to a given course, as a user.
     async register(_parent, args, { models, user }, _info) {
       const { Course, Registration } = models
@@ -435,7 +568,7 @@ const resolvers = {
           isTeacher(course, user) ||
           !canEnroll(course, now)
         ) {
-          throw new UserInputError('INVITATION_REQUEST_FAILED')
+          throw new UserInputError('INVITATION_REQUEST')
         }
 
         // Check whether there is not already a registration.